Perfect Guess
If the player guesses higher than the actual number the program displays "Lower Number Please". Similarly, if the player guesses the lower number then it will display "Higher number Please". When the player guesses the correct number the program displays the number of guesses the player used to arrive at a number.